+++ crates/wdl-format/src/v1/enum.rs
@@ -67,7 +67,7 @@ for choice in choices {
         (&choice).write(stream, config);
         if let Some(comma) = commas.next() {
             (comma).write(stream, config);
-        } else {
+        } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }
         stream.end_line();